requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
feedparser==6.0.10
tweepy==4.14.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.5
python-multipart==0.0.6
pydantic==2.5.2
aiohttp==3.9.1
